+++ kafka-junit-core/src/test/java/com/salesforce/kafka/test/KafkaTestServerTest.java
@@ -375,7 +375,7 @@ class KafkaTestServerTest {
 
         // Create SSL listener
         final BrokerListener sslListener = new SslListener()
-            .withClientAuthRequested()
+            .withClientAuthRequired()
             .withKeyStoreLocation(KafkaTestServer.class.getClassLoader().getResource("kafka.keystore.jks").getFile())
             .withKeyStorePassword("password")
             .withTrustStoreLocation(KafkaTestServer.class.getClassLoader().getResource("kafka.truststore.jks").getFile())
@@ -389,7 +389,7 @@ class KafkaTestServerTest {
 
         // Create SASL_SSL listener
         final BrokerListener saslSslListener = new SaslSslListener()
-            .withClientAuthRequested()
+            .withClientAuthRequired()
             .withUsername("kafkaclient")
             .withPassword("client-secret")
             .withKeyStoreLocation(KafkaTestServer.class.getClassLoader().getResource("kafka.keystore.jks").getFile())
@@ -417,13 +417,13 @@ class KafkaTestServerTest {
             Arguments.of(Collections.singletonList(saslPlainListener)),
 
             // Just SASL_SSL
-            Arguments.of(Collections.singletonList(saslSslListener))
+            Arguments.of(Collections.singletonList(saslSslListener)),
 
             // Combination of plain and SSL
-//            Arguments.of(listenersGroup1),
-//
-//            // Combination of SSL and SaslPlain
-//            Arguments.of(listenersGroup2)
+            Arguments.of(listenersGroup1),
+
+            // Combination of SSL and SaslPlain
+            Arguments.of(listenersGroup2)
         );
     }
 }